server/requirements.txt
# ButterflyFx Server Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# CORS
python-multipart==0.0.6

# Database (PostgreSQL)
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0

# Caching (Redis)
redis==5.0.1
hiredis==2.3.2

# Rate Limiting
slowapi==0.1.9

# Monitoring & Metrics
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0

# Utilities
python-dotenv==1.0.0

# Testing
httpx==0.26.0
pytest-asyncio==0.23.3

